Send Close when notification isn't shown
When a notification comes in and we have the screensaver or a
fullscreen application showing, we never actually show the
popup, and the application will not receive a "Close" notification.

Change this so that the close notification is the first thing
we send out after being done dealing with the requested notification.

https://bugzilla.gnome.org/show_bug.cgi?id=622021