Rework handling of image-path hints
Instead of trying to guess if the path is a URI or absolute file,
just create a GFile with g_file_new_for_commandline_arg and use
g_file_is_native to decide whether to try loading from a file.
If the resulting file object is not native, or loading from the
file fails, try loading from the icon theme.

https://bugzilla.gnome.org/show_bug.cgi?id=623508